+++ src/lib.rs
@@ -36,5 +36,5 @@ pub mod meta_examples;
 pub use crate::rotonda_store::*;
 
 // re-exports
-pub use routecore::addr;
+pub use inetnum::addr;
 pub use crossbeam_epoch::{self as epoch, Guard};